+++ src/OpenTelemetry.Instrumentation.Owin/Implementation/OwinInstrumentationActivitySource.cs
@@ -1,7 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 using System.Diagnostics;
 using System.Reflection;
 
@@ -13,9 +12,7 @@ internal static class OwinInstrumentationActivitySource
     internal static readonly string ActivitySourceName = AssemblyName.Name;
     internal static readonly string IncomingRequestActivityName = ActivitySourceName + ".IncomingRequest";
 
-    private static readonly Version Version = AssemblyName.Version;
-
-    public static ActivitySource ActivitySource { get; } = new(ActivitySourceName, Version.ToString());
+    public static ActivitySource ActivitySource { get; } = new(ActivitySourceName, SignalVersionHelper.GetVersion<OwinInstrumentationOptions>());
 
     public static OwinInstrumentationOptions? Options { get; set; }
 }