+++ src/OpenTelemetry.Instrumentation.Owin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* `ActivitySource.Version` and `Meter.Version` are set to NuGet package version.
+  ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
+
 ## 1.0.0-rc.4
 
 Released 2024-Mar-20