+++ src/OpenTelemetry.Instrumentation.Owin/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `ActivitySource.Version` and `Meter.Version` are set to NuGet package version.
   ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
+* Updated OpenTelemetry SDK to 1.8.0.
+  ([#1635](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1635))
 
 ## 1.0.0-rc.4